refactor: layout structure on external files
