+++ package.json
@@ -51,6 +51,7 @@
     "coverageProvider": "v8",
     "collectCoverageFrom": [
       "<rootDir>/src/components/pages/**/*.tsx",
+      "<rootDir>/src/components/layout/**/*.tsx",
       "!<rootDir>/src/components/pages/**/*.styles.tsx",
       "!<rootDir>/src/**/index.tsx"
     ],
