+++ src/components/layout/tabs/index.tsx
@@ -1,0 +1,1 @@
+export * from './Tabs';
